Moves map-marker models into models folder
