+++ src/models/map-marker-interface.ts
